membuat file tambah.php dan hapus.php

+++ index.php
@@ -7,33 +7,24 @@
     
 ?>
 
+<!-- HEADER -->
+<?php 
+     include_once('include/header.php')
+?>
 
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Data Makanan</title>
-
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-9ndCyUaIbzAi2FUVXJi0CjmCapSmO7SnpJef0486qhLnuZ2cdeRhO02iuK6FUUVM" crossorigin="anonymous">
-
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-icons/1.10.5/font/bootstrap-icons.min.css" integrity="sha512-ZnR2wlLbSbr8/c9AgLg3jQPAattCUImNsae6NHYnS9KrIwRdcY9DxFotXhNAKIKbAXlRnujIqUWoXXwqyFOeIQ==" crossorigin="anonymous" referrerpolicy="no-referrer" />
-
-</head>
-<body class="bg-dark text-white">
-    <div class="row text-center">
-        <div class="col-12 p-5">
-            <p class="h4"> Data Makanan Papua <i class="bi bi-egg-fried"></i> </p>
-        </div>
+<div class="row text-center">
+    <div class="col-12 p-5">
+        <p class="h4"> Data Makanan Papua <i class="bi bi-egg-fried"></i> </p>
     </div>
+</div>
 
-    <div class="container">
-        <div class="row">
-            <div class="col-12">
-                    <a href="tambah.php" class="btn btn-success" > <i class="bi bi-plus-circle"></i> Tambah Makanan </a>
+<div class="container">
+    <div class="row">
+        <div class="col-12">
+            <a href="tambah.php" class="btn btn-success"> <i class="bi bi-plus-circle"></i> Tambah Makanan </a>
 
 
-                <div class="table-responsive mt-3">
+            <div class="table-responsive mt-3">
 
                 <table class="table table-bordered table-hover">
                     <tr class="text-center">
@@ -42,6 +33,7 @@
                         <th>Harga</th>
                         <th>Warung</th>
                         <th>Alamat Warung</th>
+                        <th>Pemilik Warung</th>
                         <th> Aksi </th>
                     </tr>
                     <?php $i = 0; foreach($data as $d): ?>
@@ -51,10 +43,12 @@
                         <td>Rp. <?= number_format($d['harga_makanan'],0,',','.'); ?></td>
                         <td><?=$d  ['warung']?></td>
                         <td><?=$d ['alamat']?></td>
+                        <td><?=$d ['pemilik_warung']?></td>
                         <td>
-                            <a href="detail.php?id=<?= $d['id_makanan']?>" class="btn btn-primary"> <i class="bi bi-eye-fill"></i></a>
-                            <a href="ubah.php" class="btn btn-success"> <i class="bi bi-pencil-fill"></i></a>
-                            <a href="hapus.php" class="btn btn-danger"> <i class="bi bi-trash-fill"></i></a>
+                            <a href="detail.php?id=<?= $d['id_makanan'];?>" class="btn btn-primary"> <i
+                                    class="bi bi-eye-fill"></i></a>
+                            <a href="ubah.php?id_makanan=<?= $d['id_makanan'];?>" class="btn btn-success"> <i class="bi bi-pencil-fill"></i></a>
+                            <a href="hapus.php?id_makanan=<?= $d['id_makanan'];?>" onclick="return confirm('Apakah Anda Yakin Ingin Menghapus Datanya?');" class="btn btn-danger"> <i class="bi bi-trash-fill"></i></a>
                         </td>
                     </tr>
                     <?php endforeach ?>
@@ -65,13 +59,14 @@
 
 
 
-                </div>
             </div>
         </div>
     </div>
+</div>
 
+ <!-- CONTENT END -->
 
-
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js" integrity="sha384-geWF76RCwLtnZ8qwWowPQNguL3RmwHVBC9FhGdlKrxdiJJigb/j/68SIy3Te4Bkz" crossorigin="anonymous"></script>
-</body>
-</html>
+    <!-- footer -->
+    <?php 
+     include_once('include/footer.php')
+    ?>
